Правки после код-ревью

+++ src/main/kotlin/org/sonso/bludceapi/dto/ws/InitPayload.kt
@@ -2,6 +2,7 @@ package org.sonso.bludceapi.dto.ws
 
 import org.sonso.bludceapi.dto.ReceiptType
 import org.sonso.bludceapi.dto.TipsType
+import java.math.BigDecimal
 import java.util.*
 
 data class InitPayload(
@@ -9,7 +10,7 @@ data class InitPayload(
     val userId: UUID,
     val receiptType: ReceiptType,
     val tipsType: TipsType,
-    val amount: Double,
-    val fullAmount: Double,
+    val amount: BigDecimal,
+    val fullAmount: BigDecimal,
     val state: List<WSResponse>
 )
