+++ src/main/kotlin/org/sonso/bludceapi/dto/ws/UpdatePayload.kt
@@ -1,8 +1,10 @@
 package org.sonso.bludceapi.dto.ws
 
+import java.math.BigDecimal
+
 data class UpdatePayload(
     val type: String = "UPDATE",
-    val amount: Double,
-    val fullAmount: Double,
+    val amount: BigDecimal,
+    val fullAmount: BigDecimal,
     val state: List<WSResponse>
 )
